Control the city lights via a lights tab
We had lights in the scene, but no way to affect them. No more! The
new lights tab allows us to change the color of the different lights
in the scene, which are randomly distributed between four different
groups.

Since the lights are down in the FootprintBuildings, I decided to make
use of Godot's group system - as stated, each light is associated with
one of four groups. Changes in the Lights tab trickle up to the
MaterialColorControl, which then calls the appropriate visibility and
color set-methods.

Having messed around with this, I've made two discoveries: firstly, I
didn't really think some of these color profiles through. I need to
really think hard about what light colors I'm using in order for this
to look good.

Secondly, the shader I cooked up doesn't actually mix more than two
light sources. It seems to do a sort of A-or-B thing. That needs to be
fixed before we go ahead.