Build out a new Dot Window Light Shader
The old light shader wasn't working out - it could only tolerate one
light at a time. I need the shader to deal with multiple lights, but
it was only accepting one light at a time. So I sat down and tried to
do something better.

Firstly, I was working in VisualScript because I thought that was
easier to digest. I still think that, but this algorithm became
complex enough that I switched over to the classical shader language.
This allows for the algorithm to be more compact AND - better yet - I
can leave as many comments as I need (a LOT).

I learned a lot about writing light shaders. Maybe I'm not a master,
but I now at least have some sort of note about what's going on. For
example - "black" in a light shader is not a color, it merely denotes
the absence of light. That took a lot of trial-and-error to figure
out.

I'm actually pretty happy with the overall effect! Now I need to
modify the GUI so we can change these shader uniforms in real time...