Attempt to correct building clipping
The supposed advantage of the grow-point system is that it could
guarantee that each footprint did not impose on any other footprint.
The grow points literally stop the second they reach another point,
so there should be no overlap.

BUT! We have been getting buildings that overlap with each other all
the same. I'm fairly confident in the growth algorithm. I tested it
a lot and never had any issues. I even updated the Footprint Buildings
with an option to stop them from rotating - I was able to observe no
clipping when that was on.

If perfectly straight buildings had no problems, then the error must
lie with how we were rotating the buildings. I wasn't able to squash
all of the building clipping - perhaps my faith in the growth
algorithm was misplaced - but I was able to get a good share of it.

Firstly, I clamped the rotation between -45 and 45. I had meant to do
this originally, as I found that the rectangular nature of these
buildings means that rotating OVER +-45 produces no noticeable result.

Next, I noticed that our rotation algorithm was busted - when were
checking if the building was still in the footprint, we checked the A
point, then checked the B point three times. I did not mean to do
that. That was supposed to check A, B, C, and D. I think this is where
most of the errors were coming from.

Finally, I noticed we were rotating the whole footprint node in
addition to the buildings. The footprint's rotation was meant to be
fixed, while the building could be freely rotated. What I was doing
was effectively double rotation. Removing that fixed some more of the
rotation problems.

The clipping is now rare, but still happening. Ultra-thin buildings
are also more common now, so I might need to update the progressive
shrink algorithm with other resolution methods - perhaps a progressive
"return to zero"?