Ensure the Tabs container in Main is working
Every single child of the TabsContainer, called Tabs, was a Tab node.
I thought that was how it worked. That is not how it works. In fact,
you can put anything under the TabsContainer and it just makes it into
tabs. Anything EXCEPT a "Tab" node. Which is what I did exclusively,
and why I thought everything was so broken.

I had to go over my GUI scenes and replace the Tab nodes with the base
containers - a bit annoying, but those scenes are better organized
now.