Prototype out building beacons
A common feature of taller buildings in modern cities are beacons -
little lights sat on the edges and corners of buildings.

I wasn't entirely sure how to go about this, so I decided to play
around. Initially, I used sprites, but this caused the spawn-in time
to spike massively - it went from about 5 seconds to a full minute, if
not longer. Most likely this is because each sprite is its own thing,
possibly it's own material. It was dragging down the framerate to ~40
FPS too.

Obviously we needed to minimize the amount of materials in use, as
well as the meshes. So, I switched over to using a single quad mesh
and a single material. We'd have the material be billboarded, because
that would most closely mirror how beacons typically appear.
Internally, I was speculating how we could do multiple colors.
Initally I wanted to reuse the multicolor shader system, but I quickly
realized that would mean making all of the meshes for all of the
beacons manually. That's accetable for the buildings, but a for a
small effect.

Although the QuadMesh was good, it was severely limiting in how well
we could scale it up and down - anything over .002 looked excessive,
but Godot wouldn't let us go smaller than that. Eventually, I switched
over to using a Point Mesh, which allows us to scale based on the
material. Appropriately, I created three different materials which
will allow us three different beacons of three different patterns,
colors, and sizes.