Integrate the Beacons
After much trial and tribulation - and far more effort than it was
really worth - I have finished the beacons, complete with the GUI and
Material-Color-Control integration.

And you know what? They suck. They're not very good. I don't like the
look of them at all. It's not so much that they look bad as all of
these color schemes are based on this very simplistic appearance.
These very simplistic color schemes. Everything is impressionistic,
almost rendered down in silhouette form. And, when adding the beacons
to that, it just didn't look good. It looked distracting. Doesn't bode
very well for the other decorations, does it?

I think we'll still be fine, and here's the reasoning: the beacons
were applied to every building above a given height - that's how it
goes in real life. Building codes specify that buildings in excess of
a given height need these beacon lights. So you had this eye-grabbing
thing applied to every building of a given height, and it constantly
pulled the eyes upward. We can be much more sparing with the other
decorations - limited advertisements and such.

I think the key here is subtlety. That's why only one color scheme has
the beacons enabled. The Sunset profile has beacons so subtle that
they melt into the sky but appear brightly against the dark buildings.
That's technically the least stylized of them all, and it's the one
I'm happiest with.