Control the Sparkling effect via the GUI
Pretty standard business here - new effect, new control GUI. This
time, it's sparkles. I'm pretty happy with the GUI. I decided against
modifying the particle materials, to keep a consistent look-and feel.
We'll see if that sticks.

One one hand, I'm really happy with this effect. However, I have to
say that the shear number of configurables is starting to get out of
hand. The sparkles had 21 separate configurable entries, which broke
down into about 7 separate categories.

Our one saving grace is that I don't think we really have many
different effects left. The sparkles were, as of this moment, the only
remaining ones that really need this level of customization. The
antennae, the AC Boxes, and the advertisements - these are all things
that are simply placed. All that will change for the advertisements
are the underlying textures.