Convert to Godot 3.5
They just released Godot 3.5 a... week ago? Few days? Can't remember.
I like to use the latest-and-greatest-(and-stable), so I upgraded to
the 3.5 release.

However, the 3.5 release came with a nifty new feature: unique scene
names. This actually makes the legwork of referring to GUI nodes much
simpler, and shorter. You're free to move the nodes around as you
please, and you can use a simple '%' to refer to the node anywhere.
Excellent!