Merge together the Starfield and Stars tabs together
My GUI design has been slowly improving, and I'm starting to take a
critical eye to some of my earlier design choices. The separation
between the Starfield and the Star controls made less & less sense as
time goes on. However, the overhead of fixing that felt too great.

BUT! With the Godot 3.5 update, we can take advantage of the scene
unique names to quickly and easily rearrange the GUI. It only took me
about an hour or so to get the two GUIs merged - and I'm pretty happy
with the result!