Control the sparkle particle effect scale variation
Look, I know that the sparkles GUI is basically clogged but for some
reason I had this bee bouncing around my skull, whispering that the
users really ought to be able to customize the particle materials. At
least the scale, whispers the bee.

I caved. The bee won. I stuck the scale and scale randomness controls
in the GUI. Curse that bee, it doesn't seem to have made a difference.
Bah! I'll leave it all the same.