Remove Qodot
Qodot was originally put into this project so that we could make
custom buildings - unique buildings that broke the mold of our whole
square... thing. This was before I developed the growth-block concept
and was trying to solve how to a create a non-colliding city.

Ultimately I grew away from that concept simply because I am not such
a great engine of creativity to flawlessly produce building after
building. That's why I do procedural stuff! I ultimately replaced this
Qodot-block scheme with the current growth block system.

At this point, Qodot was just some excess weight. I've been
sentimental about project files before, but not this time. This time I
deleted the whole of Qodot, the textures directory, and the associated
blocks directory.