Bring the GrowthTest into working order
I had sorta neglected the GrowthTest because I genuinely got what I
needed out of the test. However, it's an important demonstration
piece for showing how the objects grow and collide with each other.

So I gave it a passover, just to make sure it was reflecting the
current city generation appropriately. It's back to working order now.
Good!