Rework the old tab GUI into a Toolbar
The old tab based GUI was difficult to navigate, because you had to go
from one end to the other. Maybe that's not the best way of putting
it - the point is that you had to progress through the tabs linearly.
To reach what you wanted, you necessarily had to go through stuff you
didn't want to. Plus, the linear nature meant we could only group
tabs/GUI elements together based on their neighbors. For example - the
Sky, Starfield, Moon, and Sparkles really ought to be grouped together
under some sort of environment label. With the tabs, there was no way
to do that.

So I've done a big stupid rework - and it really was a big & stupid
rework - converting the GUI into a Toolbar. The individual menus are
accessed from the Toolbar's options pop-ups, and spawn their own
individual pop-up menus.

I'm much happier with this GUI, BUT there are some problems. Godot's
windowing system is implemented using pop-ups, which means you can
really only have one up at a time. I wanted multiple windows, like in
GIMP or Photoshop or something. Also I implemented keyboard shortcuts
for some of these things but... I don't know. It looks weird and
doesn't exactly work? It was double pressing the button so I had to
implement a cooldown, and all the shortcuts say "Physical".

Whatever, good enough for now.