Delete the window generation scripts
I was really proud of the Bash texture generation script. It felt like
a good workaround for Godot not having any on-the-fly texture
generation. Sure, it was limited to Linux, but it felt like a robust
solution to the problem.

Except there was no problem. Turns out you CAN dynamically generate
textures in Godot.

Oops.

I've already prototyping how that's gonna work, so I decided to clean
up the scripts directory. It was nice while it lasted, but we've got
better ways of doing things now.