Prototype out a texture generation system
I used to think that there was no way to generate textures in Godot.
No way to paint individual pixels.

Turns out that was wrong. Like, not even a little bit correct. I'm
not sure what my initial research effort looked like, but clearly I
didn't try very hard.

The appropriate API was in Image - not only does it allow us to write
and read pixels directly, but it allows us to blit and blend one image
on top of the another image.

So, I've created a WindowGenerator class to generate these textures.
The overall algorithm is pretty much the same as the Bash scripts, but
done up in Godot. One improvement is that, since we can now paint
colors, we no longer need to separate red-green-blue images. Instead
we load a window image and generate that using it somewhat like a
mask. There is actually a mask_blit and a mask_blend but those didn't
use the mask's alpha (they treated it binary-like).

Now, to work on actually merging this with the existing colors and
stuff. That's gonna be more difficult...