Use existing constants for window generation
I realized that some of the constants we used for generating windows -
specifically, how many windows there's supposed to be - were already
stated in the GlobalRef utility script.

I've redone it now - we now calculate the texture size instead of
having it fixed. This size is dependent upon the Window Pixel Size
and the overall window count (which was set in GlobalRef).