Rework the BeaconControl scene with MCC GUI widgets
I decided to make each MCC GUI widget/node/thing it's own independent
class so we could easily spawn them in. I then decided to try out how
these custom widgets worked in one of our control scenes. I picked the
Beacon control scene to try it out.

My god, it's so easy! So fluid! So flawless! I can't believe I didn't
think of this earlier!

Currently my only concern is with the FilePathOptionButton. Before,
each control node would perform the reload only once, and update all
the relevant OptionButtons in the scene. Now they're all doing it
separate from each other, which in some case is tripling the work
load. It doesn't seem to be too much of a problem now, but it's
something to keep in mind.