Rework all of the Control GUIs
Everything is just so easy now! I went through the control nodes and
replaced them with their MCC components. And it only took me an hour!
AND THEY JUST WORK! IT USED TO TAKE ME HALF A DAY TO DO ONE AND IT
TOOK ME AN HOUR TO DO ALL OF THEM WAS I DOING WITH MY LIFE

Suffice to say, I'm very happy with how the MCC widgets have turned
out. I was able to eliminate so much boilerplate code - only one
control node still has a script, and that's just to echo out a button
signal. Now I can actually make and merge GUIs in a timely manner...