Use three different materials for buildings
The whole idea behind generating textures on the fly is to diversify
the buildings that exist in the city. However, we were only using one
texture, which meant one style of building. Ergo, we need to add more
materials to help with that diversification.

So, I whipped up three new materials and the corresponding GUI
controls. Thanks to the last commit, all of the GUI stuff was pretty
easy. The actual MCC and Profile resource updates? Not so much. It'd
be a lot easier if I could've just used JSON instead of this resource
files but Godot refuses to recgonize it! Just wish there was a better
way, that's all. Guess we're mostly over the hump now.