Use a consistent format for generated textures
Alright, so Godot supports a lot of different image formats. But
they're not like 'PNG' or 'JPEG', it's more like 'RGBE9995' and
'DXT5'. Apparently it's OpenGL formats. I don't have the capacity
right now to learn the ups-and-downs-and-ins-and-outs of all these
different formats.

But, on the other hand, I started to get worried about users adding
their own textures and breaking crashing the program. In order to
blit and blend images, they need to be the same format or...
something bad happens. It crashes? Or maybe nothing happens. The point
is that neither is good for the user.

The solution is easy then - force everything to be one format using
the Image's convert() method. From what I can tell, a RGBA8 is the
standard format for PNG, so I went with that. If there's a performance
impact on the generation, it seems negligible.