Add a filepath-based CheckBox list
The idea for our multiple texture system is to allow the user to pick
any number of window textures and then to regenerate the dynamic
textures using those chosen window textures.

The natural Godot widget for this would be the ItemList, which I tried
to work with initially. However, that widget has some problems.
Selecting an item in ItemList is a bit... unclear? It just sort of
highlights the selection. I wasn't too happy about that, which was only
compounded by the fact that everything in the ItemList must be done by
index, unlike the OptionsButton that allows you to specify by ID. This
makes it difficult to update the items on the fly, which I was trying
to do to make the selections more clear.

Eventually, I got fed up. I realized that what I really wanted was a
list of CheckBoxes. It'd actually be simple to implement; I just had
to sit down and do it.

So I did! I mostly copied code from the FilePathOptionButton. It was
a pretty easy implementation. I was tempted to make a non-file path
version, but I don't really need it right now. That, and I'd need to
concoct a scheme for automatically populating it with something OTHER
than file paths.