Add a (slightly) better dialog window
I was trying to add my new Checkbox List widget/node/scene to the
existing control dialogs when I kept running into the dialogs... just
being bad. Just straight-up awful. Why don't they size to their
children? I was going to have to manually adjust each window for the
new GUI adjustments I was going to make. And you know what? That's
just not acceptable to me.

So I created a better dialog. Well, slightly. It still has problems
and introduces some new problems but it solves the problems that I had
so... good enough? You just have to be careful with the size flags.
Setting it to "expand" can create a never-shrinking window. Or maybe
that's only with a Full Rect layout? Hard to say.