Shrink the dialog down when it first appears
I noticed that the dialog windows were appearing at some default size,
instead of the minimal size. This came down to two factors.

First, the resize was deliberately designed to stop windows from
getting too small, not from them getting too big. So there was no
shrink functionality! I've updated the _resize() function with a
shrink option so we can make it smaller if we mean to.

The other problem was not what we were doing, but when. Trying to set
the size when the pop-up was about to appear didn't work because
there's some sort of issue changing a Control node's rect_size when...
something. I'm really not sure what the condition is! Just in the
middle of some sort of processing I guess. Other people have had this
issue before, there's been some PRs submitted - the current solution
is to use yield() to deliberately delay the GUI resize to an idle
frame. Overly complex, perhaps, but acceptable.