Correct issues with generated textures
So I was going about setting up the automatic texture generation
process when I noticed that there were some things with the generated
textures were just wrong.

The first issue was a noticeable ring of alpha'd out pixels around the
windows. This seemed to happen regardless of the particular texture we
chose. I'm not sure why, but this was the result  of using blend_rect
instead of blit_rect. Where as blit is a direct copy, the blend
function blended with existing pixels at the copy site. Somehow that
was introducing some sort of errant pixels that were then getting
alpha'd out. I honestly didn't bother tracking down exactly what was
causing the issue, I was happy just changing from blend to blit.

The second issue was that, when looking at textures at odd angles, the
textures would slowly degrade into a sort of pixel soup or just
straight-up disappear. The solution there was to apply anisotropic
filtering, which helps textures maintain sharpness and definition when
the textures are at "oblique angles". To help make it look better, I
increased the anisotropic filtering level in the project's settings.
Apparently this has performance impacts, but maybe we can get away
with it since we only have three textures?