Add a Fixed MCC Options button
I realized that the user needs to be able to select the algorithm they
want to use to generate the building textures. We didn't really have a
good GUI element for that. It'd need to be something constant, and
allow the user many options but only have them pick one at a time. An
options button, in other words. But our existing MCC options button
was for loading file paths.

I made a new one. In this case, the value set is the ID of a
particular item. How we choose to decode that is up in the air.