Dynamically generate building textures at run time
As previously stated, the cities looked good but a bit samey. We had
use regenerated textures because I thought we had no other way. But
turns out, Godot allows you to modify Image resources using a variety
of methods.

So I've reworked the GUIs and scripts to support the new generation
system. After a lot of tuning and working, it seems to finally be in
working order. Except I haven't tuned any of the profiles or added an
algorithm control. The work never ends! But having played around with
it, it's certainly worth it.

I did remove some "hollow" windows. Even with the ansiotropic
filtering those ones just never ended up looking good. Just a soup of
pixels. I also noticed that some windows were "too" contigious, so I
adjusted the edges on those as well.