Control the algorithm used to generate textures at run-time
Now you wouldn't think it, but HOW we paint the shapes on the
buildings actually has a massive impact on how the buildings look. It
can really change their look-and-feel! So it was important to me that
the user be able to select which algorithm was used to generate
certain schemes.

It took some fiddling, but we can now freely switch between the
different algorithms. I'm REALLY happy with this. It has a massive
impact on the city in just the right ways.

To ensure we selected the algorithm consistently across multiple
classes, I moved the algorithm enum into GlobalRef. Although this made
some of the syntax for accessing it pretty ugly, it ensures that
everything is correct across all classes - worth it in my book. To
fill out the FixeOptionButtons we used for the algorithm select, I
just iterated over all of the members. That gave us the strings; turns
out you can use that to get the enum's value by accessing it with [].
All this enum crap is new to me, but it works and it works well!