Add attribution to the profiles
I'm a really big stickler for attribution. I like to show the heritage
of ideas, how things grow and mutate over time. I also believe it's
really important to give credit for open source projects - both
because open source projects are team efforts and to dispel notions
of singular authorship. Well... most of the code is mine right now but
someday it might not be!

It always bugged me that I couldn't appropriately credit the profiles.
I tried to make the names obvious so people would Google them, but
that wasn't enough for me. I wanted to give explanations as to where
these profiles came from and how they came to be.

To do this, I added extra fields to the profiles - an author_name
field to give author credit, and an extra_notes field for
clarifications, inspirations, and other notations. To make this
apparent to the user, I updated Toolbar with a display label. This
updates with the appropriate keys (which only change on a profile
reload).