Make the moon more dynamic using two sprites
The moon used to be a sphere. Because I was lazy. Yes, I was so lazy
that instead of making a flat circle, I made an unshaded sphere
because with the crushed perspective it would be effectively the same
thing.

And this was fine until I added those new profiles. See, I wanted to
do some weird and crazy stuff with the moon for those profiles. Make
it super big and stuff. Except, then my stupid lazy sphere was
intersecting with buildings in the city's backline. The city was being
actively eaten by the moon - and that just wasn't acceptable to me.

The solution was to replace the moon with a flat texture. I then
realized we could do multiple textures, for multiple moon designs.
Then I realized if we had two layers to the moon we could do all sorts
of effects and... I went a little overboard. So we now have a
two-layered moon and way too many textures.

Although the textures are fine, they're largely variations I generated
by hand. That makes me wonder if we could SOMEHOW create these moon
textures dynamically using a shader. That'd certainly be more
effective then what I'm doing. Still, it works well enough for now.

Also, I was getting a weird error where the stars would SOMETIMES be
in front of the moon. I lowered their material's render priority to
counteract this.