Fix the Block Test
The block test was kind of busted for a bit - I'm not sure exactly
when I broke it, but I definitely did break it. I remember now: it was
when I removed the "Building Material" from the footprint building,
which I originally did in the name of greater integration into the
Material Color Control. That broke the whole "Voronoi Color" thing we
were doing in that test.

We really do need to use the voronoi colors, because they're bright
and poppy. Helps us to distinguish the buildings and see how the
city's laid out.

So I've reintroduced a Material variable to the Footprint Building.
This time, it's an "override material". Whereas previously the
material was EXCLUSIVELY how we determined the look of the building,
the override material is a fallback. If it's there, we use it.
Otherwise, we use the materials in the MCC.

My only complaint is that, programmatically, it's not very good. If
we change the material, that should take effect instantly. I'm just
not feeling that right now. I've got bigger and better things to do.