Prevent light pop-in by increasing the building's AABB
We were getting some pop-in with the lights - like, a building would
come into frame and then a few second later some of the lights would
appear on it as though they had just loaded in. They usually had at
least one light loaded, though.

What I think was happening was that the omnilights were just taking
longer to load... somehow. In that case, I just needed to give them a
little bit more time off-screen to load in.

To do that, I added a LIGHT_SCALAR - so we can scale up radius of the
lights when included in the AABB. This causes the buildings to load in
earlier and load out later, which has removed the pop-in.