Make beacons more or less common with Occurence Ratings
A lot of the decorations face a similar problem: we don't want to have
them on every building, but since they're tied to building creation,
we can't remake them. That means we kind of have to spawn them
everywhere and then just turn them on and off. That's how it works for
lights, for beacons, and how it will work for roof boxes.

But does it need to be? I thinking about roof boxes and I realized I
wouldn't want them on every roof. I just wanted a way to cut down some
of them. That's how I came up with the idea of "occurrence" - that
each decoration would be randomly assigned and hidden according to
some arbitrary user-set value.

I decided to implement this with the Beacons, in an attempt to make
them look better. Each set of beacons is given an occurrence rating.
If they fall below a global occurrence value, they are hidden.

After playing around with it, I still don't like beacons. BUT! I think
the concept works, and is good for controlling the density of an item
in a scene. Who knows, maybe I won't even like it once roof boxes are
actually implemented.