Control the roofboxes via the GUI
Now that roofboxes are implemented, we just had to control them via
the GUI.

I updated the Rooftop Box scene to behave just like the Beacons and
did the rest of the boilerplate code to get the new controls
integrated.

However, I noticed that the way we implemented beacons was just kind
of... weird. Like I can't remember why I did it like that. And,
looking at it, I can't think of a reason WHY I did it like that. I
think I might go back and refactor all that code...