Reduce the allowed lights per object to 32
Godot has a "Max Lights Per Object" setting - it caps how many light
sources can affect a single object (i.e. a single mesh or MultiMesh).
As far as I can tell, this is mostly for performance.

See, we're basically trying to brute force "city glow" (light
pollution?) by having a frankly excessive number of lights. The idea
is that it all dissolves into a luminous technicolor soup. So I had
bumped up that setting to a maximum of 64 lights-per-object, so that
each surface could be affected by more lights.

That worked fine until we introduced the Rooftop Boxes. Each box had
five sides, and we were placing two on each rooftop. That came out to
ten sides, which sat on the existing five sides for a total of fifteen
different surfaces. In other words, my brute force deployment of boxes
has TRIPLED the potential meshes involved in a composing a building.

And the framerate was not happy. I dropped from a consistent sixty to
anywhere between 20 and 40, but only on certain profiles. I use a
gaming laptop from 2019, which probably puts me in the current
mid-to-low graphics card range. And I don't want this project to be a
graphic benchmark; it's supposed to be a fun tool anyone can use and
play with.

I had to make a choice: accessibility, or appearance? What matters to
me is that this project is more accessible, so I reduced the light
count to 32. Unfortunately you need to restart the binary to actually
see this take effect so I couldn't get a side-by-side comparison. It
FEELS fine to me, so good enough I guess. This thing's gonna be open
source anyway, if someone NEEDS to change it they can.