Create a City Light scene (and refactor some decorations)
I've been pretty busy over the past week or so, and the minimal work
in this commit has been spread out over all that time. So try to keep
with me, I'm gonna try my best to reconstruct my thought process about
what was going on here.

I replaced the OmniLights we were spawning with a custom scene, the
CityLights. Trouble is, I can't quite remember WHY I did this. I think
think the advantage was two-fold. First, it allowed us to move code
concerning the creation and updating of CityLights into a discrete
script. Secondly, having the CityLight as a separate scene allows us
greater customization if we ever want to "improve" the CityLight node.

As I was replacing the city lights, I had a long hard thought about
what constituted a "decoration" or an "effect". I went back and forth
several times but ultimately decided that a "decoration" is something
that is associated with a building and is spawned as many individual
nodes.

This "many individual nodes" thing does have an important implication:
editing a value has the potential to propagate to ALL THOSE INDIVIDUAL
NODES. What that means is we can't hook into the MCC's key_changed()
signal, because that fires for EVERY SINGLE KEY CHANGE. This results
in a massive slowdown - to get around that, we're updating these
decorations directly using group function calls. That's why CityLights
and Rooftop Boxes need their own update methods; even the beacons have
an update method that needs to get called.