Experiment with how the antennae look in a scene
Honestly, I'm really unsure about what I'm doing with these antennae.
It's like the beacons - there's a chance it'll probably be good on
some profiles but I really don't have a clear vision of what I want.

So I decided to partially implement them. They look fine, honestly.
Maybe. I CAN'T TELL.

I tell you what though - the most singularly agonizing part of
designing this class was where to set the size. See, it's going to use
three separate meshes just like the sparkles. That means that we can
use/abuse the scale to create variations in height and width.

The problem was that either the mesh needed to be set-up to be
window-cell sized, or the scale needed manipulated and multiplied so
that it was window-cell sized. And I AGONIZED over this, a lot.
Because to me, if the node is at a scale of 1, that should be
window-cell scale already. BUT the problem was that meant we needed to
size up the mesh ahead of time. I don't like to do that for resources
because it decouples the value from our length constant - and besides,
it the mesh size was getting rounded up or down and becoming
inaccurate.

Faced with bad options, I picked the least bad option: a global mesh
sized 1x1 that is scaled up and down using window-cells.

See, if this was a factory of some sort and I was gluing these
together I could just make a global mesh instead of storing it. Bah!
Whatever, this works for now. I'm not crazy about these antennae but
I'll try my best to make them work.