Add a (city) block factory
Last commit, I committed to making that building factory. And that's
still happening! Just... I was having some trouble testing it. And
most of that headache was coming from the GrowBlockifier.

Now actually the blockifier itself was sound - my chief issue has been
with everything that uses the blockifier. See, I built the blockifier
to be threaded or non-threaded. The guts were kind of exposed so you
could hook into it however you pleased. The issue was that I didn't
standardize it at all. And why would I? It's not like I needed to; all
the classes that had to interact with the blockifier were done.

Of course, coming back to that after several months, I had no idea
what was happening in these classes. There was no standardized way of
running the blockifier, so those classes that ran the blockifier were
difficult to read. Baseline functional code intermixed with building
code mixed with blockifier code - it was terrible!

The solution was to standardize how the blockifier was called - in
other words, it needed to be wrapped together into a Factory. So,
that's what I've done! And it seems to work, so far. Not what I
intended to do but it's a step in the right direction.