Don't emit signals from threads
While trying to integrate the Block Factory into my work with the
Building Factory, I ran into an odd error. Godot kept complaining
that I wasn't cleaning up my threads - yet when I tried to correct
that with a loop, the loop ran forever.

What I realized was that where my loop was being called was running
INSIDE THE THREAD! Turns out, Godot doesn't work like Qt.
Slots/functions connected to a signal will run in the thread that
emitted that signal. I actually think that's a good policy, but it
required some reworking on our part.

The _physics_process function now checks for the thread status,
cleans it up if necessary, and emits the signal. I initially wanted
to avoid this since I believe _physics_process should be for other
stuff but, now that I really think about it, I guess I'm not sure
what constitutes "other" stuff. I guess it was the precision I had
a problem with, because _physics_process gets called all the time.

Anyway, I ended up getting around that by selectively enabling and
disabling the physics processing depending on whether we have a thread
or not. Now technically it's possible to end up in a state where the
physics processing is disabled but there's a thread ongoing - but
that SHOULDN'T happen.

Hopefully.