Begin making buildings using a factory pattern
Here's the lowdown - the responsibility for the construction of
buildings and their decorations was dispersed and unclear. Some of it
was done by those individual decorations, some of it was done by the
Footprint Building, and some of it was done by the actual Grow City,
or whatever the organizing entity was.

The problem was twofold: it was difficult to make changes and it was
unclear where those changes should be made. I was losing a lot of time
because I had to meditate on where things should go, and often times
I'd change my mind the next day - effectively doubling back on what I
was doing. Even when I did make up my mind, it was difficult to
actually implement what I was trying to do because of the unclear
responsibility.

That's why I decided to try implementing the "factory" software
pattern. Or some sort of offbeat version of it. Look, I didn't
actually follow a guide I just kinda guessed. Yes that was a bad idea
but the result works SO WE'RE GOING WITH IT.

We now have the new Building Factory. The Building Factory runs a
series of steps, in order, that can be modified on the fly but default
to some global values. These steps are called "Stages" (maybe I
should've called them steps (oops)) and allow us to bind similar
functionality for the blueprint (design) and construction
(implementation) phases.

While the guts of the Building Factory are kind of complicated, it's
still much more coherent and easier to use than the Footprint
Building. Speaking of - the new Template Building is going to be our
default from now on. The idea is that we'll dynamically add
decorations onto these with different phases, but the Template
Building won't have to be bothered with all that nonsense.

To actually test out how all this is working, I have a temporary Grow
Block City (ALT), and a brand new Grow City Test. It's a lot like the
old block city test (which, incidentally, I accidentally broke), but
takes advantage of all this new tech.

Kind of melted my brain a bit, but this new foundation is going to
make our lives a lot easier going forward. Now I have to implement
building visibility management, the rest of the decoration stages, and
actually start using this in the prime-time main scene. BRING IT!!!