Handle off/onscreen building movements
So we'd whipped together this new form of building, but there was
still work to be done. The Template Buiding didn't have any of it's
visibility AABB nonsense set up!

First, I added some AABB adjustments to the Autotower build stage. Not
sure if that's exactly where it should go, but it feels about as good
as anywhere else.

Next, I upgraded the Template Building. The Template Building now has
signals for when it exits and enters a screen; it also has utility
functions for setting a building to "onscreen" (visible) or
"offscreen" (not visible) modes. Doing this via function will allow us
to expand this later as needed, and save us from disabling the
visibility notifier.

Then it was just a simple matter of hooking that up into the city.
That was pretty simple, with one caveat - I noticed that in the test
scene, changing cameras caused the city to slowly "creep" offscreen.
While that meant it was working (yay!), it was kind of ruining the
test. So, I've added a boolean to control whether we do that sort of
processing at all.